move restricted navigation classes to low level package
