+++ n8-core/src/main/kotlin/co/early/n8/NavigationState.kt
@@ -2,9 +2,10 @@
 
 package co.early.n8
 
-import co.early.n8.RestrictedNavigation.NotBackStack.IsEndNode
-import co.early.n8.RestrictedNavigation.NotBackStack.IsTabHost
+import co.early.n8.lowlevel.RestrictedNavigation.NotBackStack.IsEndNode
+import co.early.n8.lowlevel.RestrictedNavigation.NotBackStack.IsTabHost
 import co.early.n8.lowlevel.LowLevelApi
+import co.early.n8.lowlevel._notBackStack
 import co.early.n8.lowlevel.render
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
