+++ n8-core/src/main/kotlin/co/early/n8/lowlevel/RestrictedNavigation.kt
@@ -1,8 +1,10 @@
-package co.early.n8
+package co.early.n8.lowlevel
 
-import co.early.n8.RestrictedNavigation.NotBackStack
-import co.early.n8.RestrictedNavigation.NotEndNode
-import co.early.n8.lowlevel.LowLevelApi
+import co.early.n8.Navigation
+import co.early.n8.lowlevel.RestrictedNavigation.NotBackStack
+import co.early.n8.lowlevel.RestrictedNavigation.NotEndNode
+
+@Suppress("ktlint:standard:function-naming")
 
 /**
  * a Navigation can be one of 3 types (BackStack, EndNode, TabHost), but we can often further
